requirements.txt
vncorenlp
py_vncorenlp
transformers
datasets
Pillow
matplotlib
opencv-python
timm
torch
torchvision
numpy
requests
tokenizers
tqdm
accelerate
opencv-python-headless
underthesea
einops
torchmetrics
wandb
pyarrow